edit tranportation+bicycle added

+++ src/main/java/org/example/sharedmobilityfxproject/model/tranportMode/Taxi.java
@@ -3,11 +3,11 @@ package org.example.sharedmobilityfxproject.model.tranportMode;
 import org.example.sharedmobilityfxproject.model.Player;
 
 public class Taxi extends Player {
-    private double speed;
+    private int speed;
     private double carbonFootprintAmount;
     private final double carbonFootprintPerKm = 170; // Assuming this is the rate for a petrol car
 
-    public Taxi(int x, int y, int stamina, int GemCount, double speed) {
+    public Taxi(int x, int y, int stamina, int GemCount, int speed) {
         super(x, y, stamina, 10, 0, GemCount); // Assuming default speed is 10 and CO2 is 0 for initialization
         this.speed = speed;
         this.carbonFootprintAmount = 75; // Initial carbon footprint amount for Taxi
@@ -26,7 +26,7 @@ public class Taxi extends Player {
         return speed;
     }
 
-    public void setSpeed(double speed) {
+    public void setSpeed(int speed) {
         this.speed = speed;
     }
